refactor: 시간 예약 섹션 상태 로직 훅으로 분리
