+++ apps/web/src/domain/challenge/feedback/live/TimeSlotSection.tsx
@@ -1,58 +1,26 @@
 'use client';
 
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { DUMMY_MENTORS, getMentorSchedule } from '../dummy';
-import type { SelectedSlot } from './types';
+import { useTimeSlotState } from './hooks/useTimeSlotState';
 import ReservationBar from './ui/ReservationBar';
 import TimeSlotGrid from './ui/TimeSlotGrid';
 import WeekNav from './ui/WeekNav';
-import { addDays, getWeekStart } from './utils';
 
 interface Props {
   selectedMentorId: string | null;
 }
 
 const TimeSlotSection = ({ selectedMentorId }: Props) => {
-  const [weekStart, setWeekStart] = useState(() => getWeekStart(new Date()));
-  const [selectedSlot, setSelectedSlot] = useState<SelectedSlot | null>(null);
-
-  useEffect(() => {
-    setSelectedSlot(null);
-  }, [selectedMentorId]);
-
-  const mentor = useMemo(
-    () => DUMMY_MENTORS.find((m) => m.id === selectedMentorId),
-    [selectedMentorId],
-  );
-
-  const schedule = useMemo(
-    () =>
-      selectedMentorId ? getMentorSchedule(selectedMentorId, weekStart) : [],
-    [selectedMentorId, weekStart],
-  );
-
-  const handlePrev = useCallback(() => {
-    setWeekStart((prev) => addDays(prev, -7));
-    setSelectedSlot(null);
-  }, []);
-
-  const handleNext = useCallback(() => {
-    setWeekStart((prev) => addDays(prev, 7));
-    setSelectedSlot(null);
-  }, []);
-
-  const handleSlotSelect = useCallback((slot: SelectedSlot) => {
-    setSelectedSlot((prev) =>
-      prev?.date === slot.date && prev?.time === slot.time ? null : slot,
-    );
-  }, []);
-
-  const handleCancel = useCallback(() => setSelectedSlot(null), []);
-
-  const handleConfirm = useCallback(() => {
-    if (!selectedSlot) return;
-    // TODO: API 연동 시 예약 확정 로직 추가
-  }, [selectedSlot]);
+  const {
+    weekStart,
+    selectedSlot,
+    mentor,
+    schedule,
+    handlePrev,
+    handleNext,
+    handleSlotSelect,
+    handleCancel,
+    handleConfirm,
+  } = useTimeSlotState(selectedMentorId);
 
   return (
     <section>
